requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.4.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
autopep8==1.4.4
botocore==1.20.112
cairocffi==1.0.2
CairoSVG==2.4.0
certifi==2018.11.29
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
charset-normalizer==2.0.4
Click==7.0
constantly==15.1.0
cryptography==3.4.8
cssselect2==0.2.1
daphne==3.0.2
decorator==4.4.0
defusedxml==0.6.0
Django==3.2.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.8.0
django-crontab== 0.7.1
django-enumfield==2.0
django-extensions==2.1.4
django-filter==2.4.0
django-model-utils==3.1.2
django-multiselectfield==0.1.12
django-notifications-hq==1.5.0
django-sendgrid-v5==0.8.1
django-silk==3.0.4
django-storages==1.9.1
djangorestframework==3.12.4
docutils==0.14
drf-extensions==0.4.0
future==0.18.2
geographiclib==1.49
googlemaps==3.1.4
gprof2dot==2017.9.19
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.7
incremental==21.3.0
ipaddress==1.0.22
Jinja2==2.10.3
jmespath==0.10.0
jsonfield==2.0.2
lockfile==0.12.2
luigi== 2.8.2
MarkupSafe==1.1.1
mod-wsgi==4.9.0
msgpack==1.0.2
munkres==1.0.12
Pillow==8.3.1
pip==21.2.4
psycopg2==2.9.1
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==2.0.2
PyJWT==1.6.4
PyMySQL==0.9.2
pyOpenSSL==19.0.0
Pyphen==0.9.5
pyproj==3.1.0
python-daemon==2.1.2
python-dateutil==2.8.1
python-http-client==3.2.5
python-magic==0.4.24
pytz==2018.7
requests==2.21.0
s3transfer==0.4.2
sendgrid==3.6.5
sendgrid-django==4.2.0
sentry-sdk==0.19.5
service-identity==18.1.0
setuptools==57.4.0
six == 1.11.0
spawn==0.1.4
sqlparse==0.3.0
starkbank-ecdsa==1.1.1
tinycss2==1.0.2
tornado ==4.5.3
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==20.4.1
typing-extensions==3.10.0.0
urllib3==1.24.1
WeasyPrint==47
webencodings==0.5.1
xlwt==1.3.0
zope.interface==5.1.0
